Add nprocs auto option to dask-worker CLI
