+++ distributed/deploy/local.py
@@ -4,11 +4,11 @@ import math
 import warnings
 import weakref
 
-from dask.utils import factors
 from dask.system import CPU_COUNT
 import toolz
 
 from .spec import SpecCluster
+from .utils import nprocesses_nthreads
 from ..nanny import Nanny
 from ..scheduler import Scheduler
 from ..security import Security
@@ -241,34 +241,6 @@ class LocalCluster(SpecCluster):
             "The `cluster.start_worker` function has been removed. "
             "Please see the `cluster.scale` method instead."
         )
-
-
-def nprocesses_nthreads(n=CPU_COUNT):
-    """
-    The default breakdown of processes and threads for a given number of cores
-
-    Parameters
-    ----------
-    n: int
-        Number of available cores
-
-    Examples
-    --------
-    >>> nprocesses_nthreads(4)
-    (4, 1)
-    >>> nprocesses_nthreads(32)
-    (8, 4)
-
-    Returns
-    -------
-    nprocesses, nthreads
-    """
-    if n <= 4:
-        processes = n
-    else:
-        processes = min(f for f in factors(n) if f >= math.sqrt(n))
-    threads = n // processes
-    return (processes, threads)
 
 
 clusters_to_close = weakref.WeakSet()
